requirements.txt
certifi==2020.6.20
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
PyJWT==1.7.1
python-dotenv==0.10.3
pytz==2020.1
requests==2.24.0
six==1.15.0
twilio==6.45.1
urllib3==1.25.10
Werkzeug==0.16.0
